Refactor metrics.rs into modular structure for better maintainability
- Decompose large 1054-line metrics.rs file into focused modules:
  * types.rs (119 lines) - Data structures and enums
  * storage.rs (178 lines) - File I/O and persistence layer
  * tracking.rs (207 lines) - Metrics tracking logic
  * analysis.rs (148 lines) - Core metrics calculation
  * reports.rs (180 lines) - Report formatting functionality
  * performance.rs (189 lines) - Performance reporting
  * bottleneck.rs (132 lines) - Bottleneck detection
  * mod.rs (45 lines) - Public API and re-exports

- All modules are under 300 lines as required
- Maintains backward compatibility through delegation
- Clear separation of concerns between data, storage, logic, and presentation
- Improved code organization and maintainability
- No breaking changes to existing public API

Resolves #153
